✨ 四隅の浮遊 UI を画面中央上部の単一バー(TopBar)に集約
散っていた chrome を 1 本の上部中央バーに統合し、四隅の浮遊 UI を撤去。

- 新規 TopBar: ロゴ(ホーム) / テーマ / コマンドパレット / Cloud 限定
  (プレゼン・Copilot・Voice・ステータス&Follow・presence・共有) / ズーム /
  コミュニティ を横一列に集約。
- BoardIdentity・TopRightCluster・Toolbar を撤去(機能は TopBar へ移設)。
- ZoomControls / StatusBar / CommunityLink に `inline` を追加し、fixed を
  外して TopBar へ埋め込み可能に(StatusBar のメニューは下開きへ反転)。
- SidePanelToggles / CopilotPill / 接続警告バナーは従来どおり据え置き。

shape/tool 系操作は Control HUD(`` ` ``)に一本化済みで、本バーは
アプリ全体系 + Cloud 導線のみを持つ。
