+++ apps/web/src/components/board-frame/community-link.tsx
@@ -1,23 +1,31 @@
 import { useNavigate } from "react-router";
 import { I } from "../ui/index.js";
 
-/** 画面左下（StatusBar の隣）: コミュニティマップへのリンク。 */
-export function CommunityLink() {
+/**
+ * コミュニティマップへのリンク。
+ * `inline` のとき fixed をやめて親レイアウト（TopBar 等）に埋め込む。
+ */
+export function CommunityLink({ inline = false }: { inline?: boolean } = {}) {
 	const navigate = useNavigate();
 	return (
 		<button
 			type="button"
 			onClick={() => navigate("/community")}
-			className="u-surface"
+			className={inline ? undefined : "u-surface"}
 			style={{
-				position: "fixed",
-				bottom: 12,
-				left: 172,
-				zIndex: 25,
+				...(inline
+					? {}
+					: {
+							position: "fixed",
+							bottom: 12,
+							left: 172,
+							zIndex: 25,
+						}),
 				padding: "7px 11px",
 				display: "inline-flex",
 				alignItems: "center",
 				gap: 6,
+				background: "transparent",
 				border: "none",
 				color: "var(--fg-secondary)",
 				fontSize: 11.5,
