+++ apps/web/src/components/board-frame/index.ts
@@ -1,6 +1,5 @@
-export { BoardIdentity } from "./board-identity.js";
 export { CommunityLink } from "./community-link.js";
 export { CopilotPill } from "./copilot-pill.js";
 export { PresencePill } from "./presence-pill.js";
-export { TopRightCluster } from "./top-right-cluster.js";
+export { TopBar } from "./top-bar.js";
 export { ZoomControls } from "./zoom-controls.js";
