--- apps/web/src/components/toolbar/index.ts
@@ -1,1 +1,0 @@
-export { Toolbar } from "./toolbar.js";
